fix: have @earliest_report_year default to 1950
Where there are no reports, we shouldn't run `#min.to_i` on an empty array.
`[].min` returns `nil`, and `nil.to_i` returns `0`, which then breaks the date selector because that's too early a start year
NOTE: I haven't checked the 1950 date with anyone, it's just arbitrary. But it will only be used when there aren't any reports present, so not massively important